creating map- bug fixes

+++ configuration.py
@@ -156,8 +156,8 @@ class WflwConf:
     orig_of_all_test_occlusion = 736
 
     number_of_all_sample = 270956  # just images. dont count both img and lbls
-    number_of_train_sample = number_of_all_sample * 0.95  # 95 % for train
-    number_of_evaluation_sample = number_of_all_sample * 0.05  # 5% for evaluation
+    number_of_train_sample = int(number_of_all_sample * 0.95)  # 95 % for train
+    number_of_evaluation_sample = int(number_of_all_sample * 0.05) # 5% for evaluation
 
     augmentation_factor = 4  # create . image from 1
     augmentation_factor_rotate = 15  # create . image from 1
@@ -235,8 +235,8 @@ class IbugConf:
 
     '''after augmentation'''
     number_of_all_sample = 134688   # afw, train_helen, train_lfpw
-    number_of_train_sample = number_of_all_sample * 0.95  # 95 % for train
-    number_of_evaluation_sample = number_of_all_sample * 0.05  # 5% for evaluation
+    number_of_train_sample = int(number_of_all_sample * 0.95)  # 95 % for train
+    number_of_evaluation_sample = int(number_of_all_sample * 0.05)  # 5% for evaluation
 
     augmentation_factor = 4  # create . image from 1
     augmentation_factor_rotate = 20  # create . image from 1
